requirements.txt
aiofiles==23.2.1
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
certifi==2026.2.25
click==8.3.2
colorama==0.4.6
dnspython==2.8.0
email-validator==2.3.0
fastapi==0.111.0
fastapi-cli==0.0.24
filelock==3.25.2
fsspec==2026.3.0
greenlet==3.4.0
h11==0.16.0
httpcore==1.0.9
httptools==0.7.1
httpx==0.28.1
idna==3.11
intel-openmp==2021.4.0
Jinja2==3.1.6
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
mkl==2021.4.0
mpmath==1.3.0
networkx==3.6.1
numpy==1.26.4
orjson==3.11.8
pillow==10.3.0
pydantic==2.7.1
pydantic_core==2.18.2
Pygments==2.20.0
python-dotenv==1.2.2
python-multipart==0.0.9
PyYAML==6.0.3
rich==14.3.4
rich-toolkit==0.19.7
shellingham==1.5.4
SQLAlchemy==2.0.30
starlette==0.37.2
sympy==1.14.0
tbb==2021.13.1
torch==2.3.0
torchvision==0.18.0
typer==0.24.1
typing_extensions==4.15.0
ujson==5.12.0
uvicorn==0.29.0
watchfiles==1.1.1
websockets==16.0
